I've refactored amber2.cpp/.h into separate files for the different Amber parsers and have cleaned up the code and naming scheme.
Renamed classes

AmberParm => AmberPrm
AmberRst => AmberRst7

SireIO::AmberFormat => SireIO::detail::AmberFormat
