--- wrapper/IO/AmberParm.pypp.hpp
@@ -1,10 +1,0 @@
-// This file has been generated by Py++.
-
-// (C) Christopher Woods, GPL >= 2 License
-
-#ifndef AmberParm_hpp__pyplusplus_wrapper
-#define AmberParm_hpp__pyplusplus_wrapper
-
-void register_AmberParm_class();
-
-#endif//AmberParm_hpp__pyplusplus_wrapper
