+++ wrapper/IO/AmberPrm.pypp.hpp
@@ -1,0 +1,10 @@
+// This file has been generated by Py++.
+
+// (C) Christopher Woods, GPL >= 2 License
+
+#ifndef AmberPrm_hpp__pyplusplus_wrapper
+#define AmberPrm_hpp__pyplusplus_wrapper
+
+void register_AmberPrm_class();
+
+#endif//AmberPrm_hpp__pyplusplus_wrapper
