--- wrapper/IO/AmberRst.pypp.cpp
@@ -1,298 +1,0 @@
-// This file has been generated by Py++.
-
-// (C) Christopher Woods, GPL >= 2 License
-
-#include "boost/python.hpp"
-#include "AmberRst.pypp.hpp"
-
-namespace bp = boost::python;
-
-#include "SireBase/findexe.h"
-
-#include "SireBase/parallel.h"
-
-#include "SireBase/tempdir.h"
-
-#include "SireBase/unittest.h"
-
-#include "SireCAS/trigfuncs.h"
-
-#include "SireError/errors.h"
-
-#include "SireIO/errors.h"
-
-#include "SireMM/amberparams.h"
-
-#include "SireMM/atomljs.h"
-
-#include "SireMM/cljnbpairs.h"
-
-#include "SireMM/internalff.h"
-
-#include "SireMM/ljparameter.h"
-
-#include "SireMaths/maths.h"
-
-#include "SireMol/amberparameters.h"
-
-#include "SireMol/atomcharges.h"
-
-#include "SireMol/atomcoords.h"
-
-#include "SireMol/atomcutting.h"
-
-#include "SireMol/atomeditor.h"
-
-#include "SireMol/atomelements.h"
-
-#include "SireMol/atomidx.h"
-
-#include "SireMol/atommasses.h"
-
-#include "SireMol/atomvelocities.h"
-
-#include "SireMol/cgatomidx.h"
-
-#include "SireMol/connectivity.h"
-
-#include "SireMol/element.h"
-
-#include "SireMol/mgname.h"
-
-#include "SireMol/molecule.h"
-
-#include "SireMol/moleditor.h"
-
-#include "SireMol/molidx.h"
-
-#include "SireMol/reseditor.h"
-
-#include "SireMol/residuecutting.h"
-
-#include "SireMol/selector.hpp"
-
-#include "SireMove/flexibility.h"
-
-#include "SireMove/internalmove.h"
-
-#include "SireStream/datastream.h"
-
-#include "SireStream/shareddatastream.h"
-
-#include "SireSystem/system.h"
-
-#include "SireUnits/units.h"
-
-#include "SireVol/cartesian.h"
-
-#include "SireVol/periodicbox.h"
-
-#include "amber2.h"
-
-#include <QDateTime>
-
-#include <QDebug>
-
-#include <QElapsedTimer>
-
-#include <QFile>
-
-#include <QHash>
-
-#include <QRegularExpression>
-
-#include <QTextStream>
-
-#include <tuple>
-
-#include "amber2.h"
-
-SireIO::AmberRst __copy__(const SireIO::AmberRst &other){ return SireIO::AmberRst(other); }
-
-#include "Qt/qdatastream.hpp"
-
-#include "Helpers/str.hpp"
-
-void register_AmberRst_class(){
-
-    { //::SireIO::AmberRst
-        typedef bp::class_< SireIO::AmberRst, bp::bases< SireIO::MoleculeParser, SireBase::Property > > AmberRst_exposer_t;
-        AmberRst_exposer_t AmberRst_exposer = AmberRst_exposer_t( "AmberRst", "This class represents an Amber-format restartcoordinate file (ascii),\ncurrently supporting these files from Amber7 to Amber16.\n\nThe format of this file is described here;\n\nhttp:ambermd.orgformats.html\n\n(specifically the AMBER coordinaterestart file specification\n\nAuthor: Christopher Woods\n", bp::init< >("Constructor") );
-        bp::scope AmberRst_scope( AmberRst_exposer );
-        AmberRst_exposer.def( bp::init< QString const &, bp::optional< SireBase::PropertyMap const & > >(( bp::arg("filename"), bp::arg("map")=SireBase::PropertyMap() ), "Construct by parsing the passed file") );
-        AmberRst_exposer.def( bp::init< QStringList const &, bp::optional< SireBase::PropertyMap const & > >(( bp::arg("lines"), bp::arg("map")=SireBase::PropertyMap() ), "Construct by parsing the data in the passed text lines") );
-        AmberRst_exposer.def( bp::init< SireSystem::System const &, bp::optional< SireBase::PropertyMap const & > >(( bp::arg("system"), bp::arg("map")=SireBase::PropertyMap() ), "Construct by extracting the necessary data from the passed System") );
-        AmberRst_exposer.def( bp::init< SireIO::AmberRst const & >(( bp::arg("other") ), "Copy constructor") );
-        { //::SireIO::AmberRst::boxAngles
-        
-            typedef ::SireMaths::Vector ( ::SireIO::AmberRst::*boxAngles_function_type)(  ) const;
-            boxAngles_function_type boxAngles_function_value( &::SireIO::AmberRst::boxAngles );
-            
-            AmberRst_exposer.def( 
-                "boxAngles"
-                , boxAngles_function_value
-                , "Return the parsed box angles" );
-        
-        }
-        { //::SireIO::AmberRst::boxDimensions
-        
-            typedef ::SireMaths::Vector ( ::SireIO::AmberRst::*boxDimensions_function_type)(  ) const;
-            boxDimensions_function_type boxDimensions_function_value( &::SireIO::AmberRst::boxDimensions );
-            
-            AmberRst_exposer.def( 
-                "boxDimensions"
-                , boxDimensions_function_value
-                , "Return the parsed box dimensions" );
-        
-        }
-        { //::SireIO::AmberRst::coordinates
-        
-            typedef ::QVector< SireMaths::Vector > ( ::SireIO::AmberRst::*coordinates_function_type)(  ) const;
-            coordinates_function_type coordinates_function_value( &::SireIO::AmberRst::coordinates );
-            
-            AmberRst_exposer.def( 
-                "coordinates"
-                , coordinates_function_value
-                , "Return the parsed coordinate data" );
-        
-        }
-        { //::SireIO::AmberRst::formatName
-        
-            typedef ::QString ( ::SireIO::AmberRst::*formatName_function_type)(  ) const;
-            formatName_function_type formatName_function_value( &::SireIO::AmberRst::formatName );
-            
-            AmberRst_exposer.def( 
-                "formatName"
-                , formatName_function_value
-                , "" );
-        
-        }
-        { //::SireIO::AmberRst::hasVelocities
-        
-            typedef bool ( ::SireIO::AmberRst::*hasVelocities_function_type)(  ) const;
-            hasVelocities_function_type hasVelocities_function_value( &::SireIO::AmberRst::hasVelocities );
-            
-            AmberRst_exposer.def( 
-                "hasVelocities"
-                , hasVelocities_function_value
-                , "Return whether or not this restart file also provides velocities" );
-        
-        }
-        { //::SireIO::AmberRst::nAtoms
-        
-            typedef int ( ::SireIO::AmberRst::*nAtoms_function_type)(  ) const;
-            nAtoms_function_type nAtoms_function_value( &::SireIO::AmberRst::nAtoms );
-            
-            AmberRst_exposer.def( 
-                "nAtoms"
-                , nAtoms_function_value
-                , "Return the number of atoms whose coordinates are contained in this restart file" );
-        
-        }
-        AmberRst_exposer.def( bp::self != bp::self );
-        { //::SireIO::AmberRst::operator=
-        
-            typedef ::SireIO::AmberRst & ( ::SireIO::AmberRst::*assign_function_type)( ::SireIO::AmberRst const & ) ;
-            assign_function_type assign_function_value( &::SireIO::AmberRst::operator= );
-            
-            AmberRst_exposer.def( 
-                "assign"
-                , assign_function_value
-                , ( bp::arg("other") )
-                , bp::return_self< >()
-                , "" );
-        
-        }
-        AmberRst_exposer.def( bp::self == bp::self );
-        { //::SireIO::AmberRst::parse
-        
-            typedef ::SireIO::AmberRst ( *parse_function_type )( ::QString const & );
-            parse_function_type parse_function_value( &::SireIO::AmberRst::parse );
-            
-            AmberRst_exposer.def( 
-                "parse"
-                , parse_function_value
-                , ( bp::arg("filename") )
-                , "Parse from the passed file" );
-        
-        }
-        { //::SireIO::AmberRst::time
-        
-            typedef double ( ::SireIO::AmberRst::*time_function_type)(  ) const;
-            time_function_type time_function_value( &::SireIO::AmberRst::time );
-            
-            AmberRst_exposer.def( 
-                "time"
-                , time_function_value
-                , "Return the current time of the simulation from which this restart\nfile was written" );
-        
-        }
-        { //::SireIO::AmberRst::title
-        
-            typedef ::QString ( ::SireIO::AmberRst::*title_function_type)(  ) const;
-            title_function_type title_function_value( &::SireIO::AmberRst::title );
-            
-            AmberRst_exposer.def( 
-                "title"
-                , title_function_value
-                , "Return the title of the file" );
-        
-        }
-        { //::SireIO::AmberRst::toString
-        
-            typedef ::QString ( ::SireIO::AmberRst::*toString_function_type)(  ) const;
-            toString_function_type toString_function_value( &::SireIO::AmberRst::toString );
-            
-            AmberRst_exposer.def( 
-                "toString"
-                , toString_function_value
-                , "" );
-        
-        }
-        { //::SireIO::AmberRst::typeName
-        
-            typedef char const * ( *typeName_function_type )(  );
-            typeName_function_type typeName_function_value( &::SireIO::AmberRst::typeName );
-            
-            AmberRst_exposer.def( 
-                "typeName"
-                , typeName_function_value
-                , "" );
-        
-        }
-        { //::SireIO::AmberRst::velocities
-        
-            typedef ::QVector< SireMaths::Vector > ( ::SireIO::AmberRst::*velocities_function_type)(  ) const;
-            velocities_function_type velocities_function_value( &::SireIO::AmberRst::velocities );
-            
-            AmberRst_exposer.def( 
-                "velocities"
-                , velocities_function_value
-                , "Return the parsed coordinate data" );
-        
-        }
-        { //::SireIO::AmberRst::what
-        
-            typedef char const * ( ::SireIO::AmberRst::*what_function_type)(  ) const;
-            what_function_type what_function_value( &::SireIO::AmberRst::what );
-            
-            AmberRst_exposer.def( 
-                "what"
-                , what_function_value
-                , "" );
-        
-        }
-        AmberRst_exposer.staticmethod( "parse" );
-        AmberRst_exposer.staticmethod( "typeName" );
-        AmberRst_exposer.def( "__copy__", &__copy__);
-        AmberRst_exposer.def( "__deepcopy__", &__copy__);
-        AmberRst_exposer.def( "clone", &__copy__);
-        AmberRst_exposer.def( "__rlshift__", &__rlshift__QDataStream< ::SireIO::AmberRst >,
-                            bp::return_internal_reference<1, bp::with_custodian_and_ward<1,2> >() );
-        AmberRst_exposer.def( "__rrshift__", &__rrshift__QDataStream< ::SireIO::AmberRst >,
-                            bp::return_internal_reference<1, bp::with_custodian_and_ward<1,2> >() );
-        AmberRst_exposer.def( "__str__", &__str__< ::SireIO::AmberRst > );
-        AmberRst_exposer.def( "__repr__", &__str__< ::SireIO::AmberRst > );
-    }
-
-}
