--- wrapper/IO/AmberRst.pypp.hpp
@@ -1,10 +1,0 @@
-// This file has been generated by Py++.
-
-// (C) Christopher Woods, GPL >= 2 License
-
-#ifndef AmberRst_hpp__pyplusplus_wrapper
-#define AmberRst_hpp__pyplusplus_wrapper
-
-void register_AmberRst_class();
-
-#endif//AmberRst_hpp__pyplusplus_wrapper
