+++ wrapper/IO/AmberRst7.pypp.hpp
@@ -1,0 +1,10 @@
+// This file has been generated by Py++.
+
+// (C) Christopher Woods, GPL >= 2 License
+
+#ifndef AmberRst7_hpp__pyplusplus_wrapper
+#define AmberRst7_hpp__pyplusplus_wrapper
+
+void register_AmberRst7_class();
+
+#endif//AmberRst7_hpp__pyplusplus_wrapper
