+++ wrapper/IO/CMakeAutogenFile.txt
@@ -1,8 +1,8 @@
 # WARNING - AUTOGENERATED FILE - CONTENTS WILL BE OVERWRITTEN!
 set ( PYPP_SOURCES
        Amber.pypp.cpp
-       AmberParm.pypp.cpp
-       AmberRst.pypp.cpp
+       AmberPrm.pypp.cpp
+       AmberRst7.pypp.cpp
        Cube.pypp.cpp
        FlexibilityLibrary.pypp.cpp
        FlexibilityTemplate.pypp.cpp
