+++ wrapper/IO/SireIO_properties.cpp
@@ -4,16 +4,6 @@
 #include "Base/convertproperty.hpp"
 #include "SireIO_properties.h"
 
-#include "SireError/errors.h"
-#include "SireMol/cuttingfunction.h"
-#include "SireMol/molecule.h"
-#include "SireMol/molidx.h"
-#include "SireMol/mover.hpp"
-#include "SireStream/datastream.h"
-#include "iobase.h"
-#include <QDebug>
-#include <QFile>
-#include "iobase.h"
 #include "SireBase/booleanproperty.h"
 #include "SireBase/parallel.h"
 #include "SireError/errors.h"
@@ -29,8 +19,18 @@
 #include <QMutex>
 #include <QTextStream>
 #include "moleculeparser.h"
+#include "SireError/errors.h"
+#include "SireMol/cuttingfunction.h"
+#include "SireMol/molecule.h"
+#include "SireMol/molidx.h"
+#include "SireMol/mover.hpp"
+#include "SireStream/datastream.h"
+#include "iobase.h"
+#include <QDebug>
+#include <QFile>
+#include "iobase.h"
 void register_SireIO_properties()
 {
-    register_property_container< SireIO::IOPtr, SireIO::IOBase >();
     register_property_container< SireIO::MoleculeParserPtr, SireIO::MoleculeParser >();
+    register_property_container< SireIO::IOPtr, SireIO::IOBase >();
 }
